requirements.txt
scikit-image
matplotlib
nibabel
tk
Pillow
scipy
antspyx